added restrictions to post access from users, only users who created posts can access it

+++ app/routers/post.py
@@ -10,8 +10,8 @@ router = APIRouter(prefix = "/posts", tags = ['posts'])
 
 
 @router.post("/", status_code = status.HTTP_201_CREATED, response_model = schemas.Post)
-async def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db), get_current_user: int = Depends(oauth2.get_current_user)):
-    query = models.Post(**post.model_dump())
+async def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+    query = models.Post(user_id = current_user.id, **post.model_dump())
     db.add(query)
     db.commit()
     db.refresh(query)
@@ -21,8 +21,8 @@ async def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db), 
     return query
 
 @router.get("/", response_model = List[schemas.Post])
-async def get_all_posts(db: Session = Depends(get_db)):
-    query = db.query(models.Post).all()
+async def get_all_posts(db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+    query = db.query(models.Post).filter(models.Post.user_id == current_user.id).all()
     return query
 
 
@@ -37,12 +37,14 @@ async def get(id: int, db: Session = Depends(get_db)):
 
 
 @router.delete("/{id}")
-def delete(id:int, db: Session = Depends(get_db), get_current_user: int = Depends(oauth2.get_current_user)):
+def delete(id:int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     query = db.query(models.Post).filter(models.Post.id == id)
     
     if query.first() is None:
         raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, 
                             detail = f"post with id: {id} was not found") 
+    if query.first().user_id != current_user.id:
+        raise HTTPException(status_code= status.HTTP_403_FORBIDDEN, detail=" Not Authorised to perform requested action")
     query.delete(synchronize_session = False)
     db.commit()
     return Response(status_code = status.HTTP_204_NO_CONTENT)
@@ -50,12 +52,14 @@ def delete(id:int, db: Session = Depends(get_db), get_current_user: int = Depend
 
 
 @router.put("/{id}", status_code = status.HTTP_200_OK)
-def update(id: int, post: schemas.PostCreate, db: Session = Depends(get_db), get_current_user: int = Depends(oauth2.get_current_user)):
+def update(id: int, post: schemas.PostCreate, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     query = db.query(models.Post).filter(models.Post.id == id)
 
     if query.first() is None:
         raise HTTPException(status_code = status.HTTP_404_NOT_FOUND, 
                             detail = f"post with id: {id} was not found") 
+    if query.first().user_id != current_user.id:
+        raise HTTPException(status_code= status.HTTP_403_FORBIDDEN, detail=" Not Authorised to perform requested action")
     query.update(post.model_dump(), synchronize_session = False)
     db.commit()
     return query.first()
